Link's GetPos and GetRot improvements
Functions being modified to being able to get Position and Rotation of the link for the lobejct which will be added to main one.
TODO: some additional models still have wrong placement regarding of gained Vec3 values and some models still not show up. It seems that issue was not in the incorrect coordinates value that being gotten by functions