Multi cannon spawn fix for seas_great_cannon
Fixed issue, when during each time saving the map, game was generating a new object "seas_great_cannon" even if the previous one was already exist.

To fix issues on the maps, required to remove duplicates on all maps which were saved after that issue appeared. Better to delete whole unit and place on map again.