Pirates GfxPrefix Flag Save Fix
Because of multiple calls of ResetFlag(0 function by CTransportObj logic, after loading up Save it removes Linked flag, but the source code of GfxPrefix logic is not working well after that and not adding flag back to already existing objects on map. To prevent that, the flag addding enforced through level CheckLevelFlag() function like it was done in previous Commits, in order to fix the issue in all cases.